building the modules
